Move to new utils file
